Comment and Models

+++ Models/History.cs
@@ -8,7 +8,7 @@ namespace GypooWebAPI.Models
         [BsonRepresentation(BsonType.ObjectId)]
         public string? _id { get; set; } = null!; // id ของประวัติให้ที่พัก
         public int? index { get; set; } = 0;
-        public float price { get; set; } = 0;
+        public double price { get; set; } = 0;
         public List<string> picture { get; set; } = new List<string>(); // Hotel Picture
         public string name { get; set; } = null!; // Hotel name
         public string type { get; set; } = null!; // type room
